backend/requirements.txt
# ===== FastAPI Backend =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ===== Speech-to-Text (Whisper) =====
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
soundfile==0.12.1
pydub==0.25.1

# ===== NLP & Sentiment Analysis =====
transformers==4.36.2
spacy==3.7.2
nltk==3.8.1

# ===== Database =====
sqlalchemy==2.0.25
aiosqlite==0.19.0

# ===== MLOps =====
mlflow==2.9.2

# ===== Utilities =====
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.26.3
pandas==2.1.4

# ===== Testing =====
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
